--- src/DataArray.php
@@ -1,125 +1,0 @@
-<?php
-
-// +----------------------------------------------------------------------
-// | 宝塔ThinkPhP6扩展包
-// +----------------------------------------------------------------------
-// | 版权所有 2017~2020 [ https://www.dtapp.net ]
-// +----------------------------------------------------------------------
-// | 官方网站: https://gitee.com/dtapps/think-bt
-// +----------------------------------------------------------------------
-// | 开源协议 ( https://mit-license.org )
-// +----------------------------------------------------------------------
-// | gitee 仓库地址 ：https://gitee.com/dtapps/think-bt
-// | github 仓库地址 ：https://github.com/dtapps/think-bt
-// | Packagist 地址 ：https://packagist.org/packages/liguangchun/think-bt
-// +----------------------------------------------------------------------
-
-namespace DtApp\Think\Bt;
-
-use ArrayAccess;
-
-/**
- * Class DataArray
- * @package DtApp\Think\Bt
- */
-class DataArray implements ArrayAccess
-{
-    /**
-     * 当前配置值
-     * @var array
-     */
-    private $config = [];
-
-    /**
-     * DataArray constructor.
-     * @param array $options
-     */
-    public function __construct(array $options)
-    {
-        $this->config = $options;
-    }
-
-    /**
-     * 设置配置项值
-     * @param $offset
-     * @param $value
-     */
-    public function set($offset, $value)
-    {
-        $this->offsetSet($offset, $value);
-    }
-
-    /**
-     * 获取配置项参数
-     * @param null $offset
-     * @return mixed|void
-     */
-    public function get($offset = null)
-    {
-        return $this->offsetGet($offset);
-    }
-
-    /**
-     * 合并数据到对象
-     * @param array $data 需要合并的数据
-     * @param bool $append 是否追加数据
-     * @return array
-     */
-    public function merge(array $data, $append = false)
-    {
-        if ($append) {
-            return $this->config = array_merge($this->config, $data);
-        }
-        return array_merge($this->config, $data);
-    }
-
-    /**
-     * 设置配置项值
-     * @param string $offset
-     * @param string|array|null|integer $value
-     */
-    public function offsetSet($offset, $value)
-    {
-        if (is_null($offset)) {
-            $this->config[] = $value;
-        } else {
-            $this->config[$offset] = $value;
-        }
-    }
-
-    /**
-     * 判断配置Key是否存在
-     * @param string $offset
-     * @return bool
-     */
-    public function offsetExists($offset)
-    {
-        return isset($this->config[$offset]);
-    }
-
-    /**
-     * 清理配置项
-     * @param string|null $offset
-     */
-    public function offsetUnset($offset = null)
-    {
-        if (is_null($offset)) {
-            $this->config = [];
-        } else {
-            unset($this->config[$offset]);
-        }
-    }
-
-    /**
-     * 获取配置项参数
-     * @param string|null $offset
-     * @return array|string|null
-     */
-    public function offsetGet($offset = null)
-    {
-        if (is_null($offset)) {
-            return $this->config;
-        }
-        return isset($this->config[$offset]) ? $this->config[$offset] : null;
-    }
-}
